# Run LFS hooks after checking if hooks are skipped