--- a/internal/lefthook/runner/runner.go
+++ b/internal/lefthook/runner/runner.go
@@ -90,13 +90,13 @@ type executable interface {
 func (r *Runner) RunAll(ctx context.Context, sourceDirs []string) ([]Result, error) {
 	results := make([]Result, 0, len(r.Hook.Commands)+len(r.Hook.Scripts))
 
-	if err := r.runLFSHook(ctx); err != nil {
-		return results, err
-	}
-
 	if r.Hook.DoSkip(r.Repo.State()) {
 		r.logSkip(r.HookName, "hook setting")
 		return results, nil
+	}
+
+	if err := r.runLFSHook(ctx); err != nil {
+		return results, err
 	}
 
 	if !r.DisableTTY && !r.Hook.Follow {