@mrexox 2024-09-20
Hey! Thank you for creating this PR. Are you sure that LSF should be skipped when lefthook hooks are skipped? LFS seems to be like something independent from lefthook and is supposed to be executed on every pre-push hook regardless lefthook configuration.

Could you share your case, so we can discuss this?