@zghorn 2024-09-25
FWIW, I've been using @zachahn fork for this past week with no incident. The speedup is noticeable